FerreMas-main/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
stripe
pydantic